# docs/swagger.yaml
basePath: /api
definitions:
...
  structs.ResidentialArea:
    properties:
      district:
        properties:
          code:
            type: string
          id:
            type: integer
          name:
            type: string
        type: object
      region:
        type: string
      residential_areas:
        items:
          type: string
        type: array
    type: object
  structs.Village:
    properties:
      district:
        type: string
      id:
        type: integer
      village_name:
        type: string
    type: object
  structs.Ward:
    properties:
      constituency:
        type: string
      district_code:
        type: string
      district_name:
        type: string
      id:
        type: integer
      region:
        type: string
      wards:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: With this API, developers can easily access and retrieve data related
    to these administrative divisions, enabling them to build location-based applications,
    conduct geospatial analysis, or enhance existing systems with accurate and up-to-date
    information about Malawi's geographic structure.
  title: Malawi Country Data API
  version: 1.0.2
paths:
...
  /ResidentialAreas/Search/{search}:
    get:
      consumes:
      - application/json
      description: Searches though Residential Area records
      parameters:
      - description: Residential Area
        in: path
        name: search
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ResidentialArea'
      summary: Searches for Residential Areas
      tags:
      - Residential Area
...
  /Villages/District/{search}:
    get:
      consumes:
      - application/json
      description: Gets all the villages in district
      parameters:
      - description: District
        in: path
        name: district
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Village'
      summary: Gets all the country villages in a district
      tags:
      - villages
  /Villages/Search/{search}:
    get:
      consumes:
      - application/json
      description: Gets all the villages in district
      parameters:
      - description: village
        in: path
        name: village
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Village'
      summary: Searches village name
      tags:
      - villages
...
  /Wards/District/{search}:
    get:
      consumes:
      - application/json
      description: Gets all wards by district name
      parameters:
      - description: District name
        in: path
        name: search
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Ward'
      summary: Searches wards by district
      tags:
      - Wards
...
swagger: "2.0"